revert(tui): remove model selector from autopilot startup
Remove the model dropdown that was added - it's not working correctly.
The ClaudeModel enum is kept in models.py for future use.
