+++ src/shotgun/agents/autopilot/autopilot_orchestrator.py
@@ -69,10 +69,6 @@ class AutopilotConfig(BaseModel):
         default=MAX_STAGE_ITERATIONS,
         description="Maximum iterations per stage before giving up",
     )
-    model: str | None = Field(
-        default=None,
-        description="Claude model to use (None = Claude Code default)",
-    )
 
 
 class AutopilotOrchestrator:
@@ -609,7 +605,6 @@ class AutopilotOrchestrator:
         """
         subprocess_config = ClaudeSubprocessConfig(
             working_directory=self.config.working_directory,
-            model=self.config.model,
         )
         self._claude = ClaudeSubprocess(subprocess_config)
